+++ form-service/src/main/java/com/sougata/form_service/dto/form/FormResponseDto.java
@@ -15,7 +15,7 @@ public record FormResponseDto(
         Boolean acceptingResponse,
         String notAcceptingResponseMessage,
         Instant stopAcceptingResponseOn,
-        Long stopAcceptingResponseAfterResponse,
+        Integer stopAcceptingResponseAfterResponse,
         Instant lastOpenedOn,
         List<QuestionRes> questions
 ) {
